lesson5/vaia-l5-model-comparison/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
numpy==1.26.3
pandas==2.1.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
